cmd/k8s-operator: reconcile a single service at a time

+++ cmd/k8s-operator/ingress-for-pg.go
@@ -119,6 +119,8 @@ func (a *IngressPGReconciler) maybeProvision(ctx context.Context, hostname strin
 		logger.Infof("[unexpected] no ProxyGroup annotation, skipping VIPService provisioning")
 		return nil
 	}
+	logger = logger.With("ProxyGroup", pgName)
+
 	pg := &tsapi.ProxyGroup{}
 	if err := a.Get(ctx, client.ObjectKey{Name: pgName}, pg); err != nil {
 		if apierrors.IsNotFound(err) {
@@ -265,11 +267,7 @@ func (a *IngressPGReconciler) maybeProvision(ctx context.Context, hostname strin
 
 	// 5. Update tailscaled's AdvertiseServices config, which should add the VIPService
 	// IPs to the ProxyGroup Pods' AllowedIPs in the next netmap update if approved.
-	ingList := &networkingv1.IngressList{}
-	if err := a.List(ctx, ingList); err != nil {
-		return fmt.Errorf("failed to list Ingresses: %w", err)
-	}
-	if err = a.updateTailscaledConfigWithServices(ctx, pg.Name, ingList, logger); err != nil {
+	if err = a.maybeUpdateAdvertiseServicesConfig(ctx, pg.Name, serviceName, true, logger); err != nil {
 		return fmt.Errorf("failed to update tailscaled config: %w", err)
 	}
 
@@ -299,9 +297,9 @@ func (a *IngressPGReconciler) maybeProvision(ctx context.Context, hostname strin
 // maybeCleanupProxyGroup ensures that if an Ingress hostname has changed, any VIPService resources created for the
 // Ingress' ProxyGroup corresponding to the old hostname are cleaned up. A run of this function will ensure that any
 // VIPServices that are associated with the provided ProxyGroup and no longer owned by an Ingress are cleaned up.
-func (a *IngressPGReconciler) maybeCleanupProxyGroup(ctx context.Context, proxyGroupName string, logger *zap.SugaredLogger) error {
+func (a *IngressPGReconciler) maybeCleanupProxyGroup(ctx context.Context, pgName string, logger *zap.SugaredLogger) error {
 	// Get serve config for the ProxyGroup
-	cm, cfg, err := a.proxyGroupServeConfig(ctx, proxyGroupName)
+	cm, cfg, err := a.proxyGroupServeConfig(ctx, pgName)
 	if err != nil {
 		return fmt.Errorf("getting serve config: %w", err)
 	}
@@ -328,17 +326,13 @@ func (a *IngressPGReconciler) maybeCleanupProxyGroup(ctx context.Context, proxyG
 
 		if !found {
 			logger.Infof("VIPService %q is not owned by any Ingress, cleaning up", vipServiceName)
-			svc, err := a.getVIPService(ctx, vipServiceName, logger)
+
+			// Delete the VIPService from control if necessary.
+			svc, exists, err := a.tryGetVIPService(ctx, vipServiceName)
 			if err != nil {
-				errResp := &tailscale.ErrResponse{}
-				if errors.As(err, &errResp) && errResp.Status == http.StatusNotFound {
-					delete(cfg.Services, vipServiceName)
-					serveConfigChanged = true
-					continue
-				}
 				return err
 			}
-			if isVIPServiceForAnyIngress(svc) {
+			if exists && isVIPServiceForAnyIngress(svc) {
 				logger.Infof("cleaning up orphaned VIPService %q", vipServiceName)
 				if err := a.tsClient.deleteVIPService(ctx, vipServiceName); err != nil {
 					errResp := &tailscale.ErrResponse{}
@@ -346,6 +340,11 @@ func (a *IngressPGReconciler) maybeCleanupProxyGroup(ctx context.Context, proxyG
 						return fmt.Errorf("deleting VIPService %q: %w", vipServiceName, err)
 					}
 				}
+			}
+
+			// Make sure the VIPService is not advertised in tailscaled or serve config.
+			if err = a.maybeUpdateAdvertiseServicesConfig(ctx, pgName, vipServiceName, false, logger); err != nil {
+				return fmt.Errorf("failed to update tailscaled config services: %w", err)
 			}
 			delete(cfg.Services, vipServiceName)
 			serveConfigChanged = true
@@ -361,11 +360,6 @@ func (a *IngressPGReconciler) maybeCleanupProxyGroup(ctx context.Context, proxyG
 		if err := a.Update(ctx, cm); err != nil {
 			return fmt.Errorf("updating serve config: %w", err)
 		}
-	}
-
-	// 3. Unadvertise the VIPService in tailscaled config.
-	if err = a.updateTailscaledConfigWithServices(ctx, proxyGroupName, ingList, logger); err != nil {
-		return fmt.Errorf("failed to update tailscaled config services: %w", err)
 	}
 
 	return nil
@@ -407,11 +401,7 @@ func (a *IngressPGReconciler) maybeCleanup(ctx context.Context, hostname string,
 	}
 
 	// 3. Unadvertise the VIPService in tailscaled config.
-	ingList := &networkingv1.IngressList{}
-	if err := a.List(ctx, ingList); err != nil {
-		return fmt.Errorf("failed to list Ingresses: %w", err)
-	}
-	if err = a.updateTailscaledConfigWithServices(ctx, pg, ingList, logger); err != nil {
+	if err = a.maybeUpdateAdvertiseServicesConfig(ctx, pg, serviceName, false, logger); err != nil {
 		return fmt.Errorf("failed to update tailscaled config services: %w", err)
 	}
 
@@ -503,16 +493,17 @@ func shouldExpose(ing *networkingv1.Ingress) bool {
 	return isTSIngress && pgAnnot != ""
 }
 
-func (a *IngressPGReconciler) getVIPService(ctx context.Context, name tailcfg.ServiceName, logger *zap.SugaredLogger) (*VIPService, error) {
+func (a *IngressPGReconciler) tryGetVIPService(ctx context.Context, name tailcfg.ServiceName) (_ *VIPService, exists bool, _ error) {
 	svc, err := a.tsClient.getVIPService(ctx, name)
 	if err != nil {
 		errResp := &tailscale.ErrResponse{}
-		if ok := errors.As(err, errResp); ok && errResp.Status != http.StatusNotFound {
-			logger.Infof("error getting VIPService %q: %v", name, err)
-			return nil, fmt.Errorf("error getting VIPService %q: %w", name, err)
+		if ok := errors.As(err, errResp); ok && errResp.Status == http.StatusNotFound {
+			return nil, false, nil
 		}
+
+		return nil, false, fmt.Errorf("error getting VIPService %q: %w", name, err)
 	}
-	return svc, nil
+	return svc, true, nil
 }
 
 func isVIPServiceForIngress(svc *VIPService, ing *networkingv1.Ingress) bool {
@@ -576,12 +567,14 @@ func (a *IngressPGReconciler) validateIngress(ing *networkingv1.Ingress, pg *tsa
 
 // deleteVIPServiceIfExists attempts to delete the VIPService if it exists and is owned by the given Ingress.
 func (a *IngressPGReconciler) deleteVIPServiceIfExists(ctx context.Context, name tailcfg.ServiceName, ing *networkingv1.Ingress, logger *zap.SugaredLogger) error {
-	svc, err := a.getVIPService(ctx, name, logger)
+	svc, exists, err := a.tryGetVIPService(ctx, name)
 	if err != nil {
 		return fmt.Errorf("error getting VIPService: %w", err)
 	}
+	if !exists {
+		return nil
+	}
 
-	// isVIPServiceForIngress handles nil svc, so we don't need to check it here
 	if !isVIPServiceForIngress(svc, ing) {
 		return nil
 	}
@@ -593,15 +586,8 @@ func (a *IngressPGReconciler) deleteVIPServiceIfExists(ctx context.Context, name
 	return nil
 }
 
-func (a *IngressPGReconciler) updateTailscaledConfigWithServices(ctx context.Context, pgName string, ingList *networkingv1.IngressList, logger *zap.SugaredLogger) error {
-	var servicesToAdvertise []string
-	for _, i := range ingList.Items {
-		if i.DeletionTimestamp.IsZero() && shouldExpose(&i) {
-			// TODO(tomhjp): Also check that it's ready.
-			servicesToAdvertise = append(servicesToAdvertise, hostnameForIngress(&i))
-		}
-	}
-	logger.Debugf("setting services to advertise: %q", servicesToAdvertise)
+func (a *IngressPGReconciler) maybeUpdateAdvertiseServicesConfig(ctx context.Context, pgName string, serviceName tailcfg.ServiceName, shouldBeAdvertised bool, logger *zap.SugaredLogger) (err error) {
+	logger.Debugf("will advertise %q: %v", serviceName, shouldBeAdvertised)
 
 	// Get all config Secrets for this ProxyGroup.
 	secrets := &corev1.SecretList{}
@@ -610,14 +596,27 @@ func (a *IngressPGReconciler) updateTailscaledConfigWithServices(ctx context.Con
 	}
 
 	for _, secret := range secrets.Items {
+		var updated bool
 		for fileName, confB := range secret.Data {
 			var conf ipn.ConfigVAlpha
 			if err := json.Unmarshal(confB, &conf); err != nil {
 				return fmt.Errorf("error unmarshalling ProxyGroup config: %w", err)
 			}
 
-			// Set the new data.
-			conf.AdvertiseServices = servicesToAdvertise
+			// Update the services to advertise if required.
+			idx := slices.Index(conf.AdvertiseServices, serviceName.WithoutPrefix())
+			isAdvertised := idx >= 0
+			switch {
+			case isAdvertised == shouldBeAdvertised:
+				// Already up to date.
+				continue
+			case isAdvertised:
+				// Needs to be removed.
+				conf.AdvertiseServices = slices.Delete(conf.AdvertiseServices, idx, idx+1)
+			case shouldBeAdvertised:
+				// Needs to be added.
+				conf.AdvertiseServices = append(conf.AdvertiseServices, serviceName.WithoutPrefix())
+			}
 
 			// Update the Secret.
 			confB, err := json.Marshal(conf)
@@ -625,6 +624,10 @@ func (a *IngressPGReconciler) updateTailscaledConfigWithServices(ctx context.Con
 				return fmt.Errorf("error marshalling ProxyGroup config: %w", err)
 			}
 			mak.Set(&secret.Data, fileName, confB)
+			updated = true
+		}
+
+		if updated {
 			if err := a.Update(ctx, &secret); err != nil {
 				return fmt.Errorf("error updating ProxyGroup config Secret: %w", err)
 			}
